changed folder structure
I was able to add the favicon. I played around with a navigation bar, though this is not what we shall put on the site. We shall have it hidden in a menu. I also made a mistake of making these trials in the main branch. I should not do this the next time since we have to avoid it. 1-09-2023 9:42