applied css to the logo and menu icon
I put the logo and the menu icon in their respective places. I need to align them, or their div containers. The next steps are to finish up the upper section of the home page before I touch the other parts. When done with these, we shall apply functionality to them.